storage: Make "installation applied" status generic
For apps-only updates, the "installation applied" status indicates that
new apps have been installed but not yet started. The existing
status value, Install applied, waiting for reboot, is confusing in this
case because no reboot is needed. To make the status message appropriate
for both OSTree and apps-only updates, we make it more generic.

Signed-off-by: Mike Sul <mike.sul@foundries.io>